+++ packages/react-codemirror/src/e2e_tests/syntaxValidation.spec.tsx
@@ -1,3 +1,4 @@
+import { testData } from '@neo4j-cypher/language-support';
 import { expect, test } from '@playwright/experimental-ct-react';
 import { CypherEditor } from '../CypherEditor';
 import { CypherEditorPage } from './e2eUtils';
@@ -86,6 +87,20 @@ test('Errors for multiline undefined labels are highlighted correctly', async ({
   await editorPage.checkWarningMessage('Bar`', expectedMsg);
 });
 
+test('Semantic errors work in firefox', async ({
+  browserName,
+  page,
+  mount,
+}) => {
+  test.skip(browserName !== 'firefox');
+  const editorPage = new CypherEditorPage(page);
+  const query = 'MATCH (n:OperationalPoint)--(m:OperationalPoint) RETURN s,m,n';
+
+  await mount(<CypherEditor value={query} schema={testData.mockSchema} />);
+
+  await editorPage.checkErrorMessage('s,m,n', 'Variable `s` not defined');
+});
+
 test('Semantic errors are surfaced when there are no syntactic errors', async ({
   page,
   mount,
@@ -153,4 +168,35 @@ test('Validation errors are correctly overlapped', async ({ page, mount }) => {
     '-1',
     "Invalid input. '-1' is not a valid value. Must be a positive integer",
   );
+});
+
+test('Strikethroughs are shown for deprecated functions', async ({ page, mount }) => {
+  const editorPage = new CypherEditorPage(page);
+  const query = `RETURN id()`;
+
+  await mount(<CypherEditor value={query} schema={testData.mockSchema} />);
+  await expect(editorPage.page.locator('.cm-deprecated-element').last()).toBeVisible(
+    { timeout: 3000 },
+  );
+  await editorPage.checkWarningMessage(
+    'id',
+    "Function id is deprecated.",
+  );
+
+});
+
+test('Strikethroughs are shown for deprecated procedures', async ({ page, mount }) => {
+  const editorPage = new CypherEditorPage(page);
+  const query = `CALL apoc.create.uuids()`;
+
+  await mount(<CypherEditor value={query} schema={testData.mockSchema} />);
+  await expect(editorPage.page.locator('.cm-deprecated-element').last()).toBeVisible(
+    { timeout: 3000 },
+  );
+
+  await editorPage.checkWarningMessage(
+    'apoc.create.uuids',
+    "Procedure apoc.create.uuids is deprecated.",
+  );
+
 });